Fixed bugs, changed syntax

+++ .localhistory/CriPakTools/1425264263$Program.cs
@@ -150,25 +150,37 @@ namespace CriPakTools
                 //    if (doTheThing) selectedEntries.Add(totalEntries[i]);
                 //}
 
-                int consoleLine = Console.CursorTop;
-                int numReplaced = 0;
+                List<FileEntry> completedEntryReplacements = new List<FileEntry>();
+                int previousLine = Console.CursorTop;
 
                 for (int i = 0; i < totalEntries.Count; i++)
                 {
-
                     if (i > 0)
                     {
-                        int numLnClr = Console.CursorTop - consoleLine;
-                        Console.SetCursorPosition(0, consoleLine);
+                        ConsoleColor prevColor = Console.ForegroundColor;
+                        Console.ForegroundColor = ConsoleColor.Blue;
+                        Console.SetCursorPosition(0, previousLine - completedEntryReplacements.Count);
 
-                        Console.WriteLine(new string(' ', Console.WindowWidth * numLnClr));
+                        for (int j = 0; j < (Console.CursorTop - previousLine) + completedEntryReplacements.Count; i++)
+                        {
+                            ClearCurrentConsoleLine();
+                            Console.WriteLine(new string(' ', Console.WindowWidth));
+                        }
 
-                        Console.SetCursorPosition(0, consoleLine);
+                        Console.SetCursorPosition(0, previousLine - completedEntryReplacements.Count);
+
+                        for (int j = 0; j < completedEntryReplacements.Count; j++)
+                        {                            
+                            Console.WriteLine("Replaced file " + (i + 1) + " of " + replaceNames.Count + ": \"" + ((completedEntryReplacements[j].DirName != null) ? (string)completedEntryReplacements[j].DirName + "/" : "") + completedEntryReplacements[j].FileName.ToString().Trim() + "\".");
+                        }
+                        Console.ForegroundColor = prevColor;
                     }
 
-                    Console.WriteLine("Updating file "+ (i + 1) + " of " + totalEntries.Count + ": \"" + ((totalEntries[i].DirName != null) ? (string)totalEntries[i].DirName + "/" : "") + totalEntries[i].FileName.ToString().Trim() + "\"...");
-                    Console.WriteLine(new string(' ', Console.WindowWidth));  
+                    previousLine = Console.CursorTop;
+                    Console.WriteLine("Checking file "+ (i + 1) + " of " + totalEntries.Count + ": \"" + ((totalEntries[i].DirName != null) ? (string)totalEntries[i].DirName + "/" : "") + totalEntries[i].FileName.ToString().Trim() + "\"...");
+                    Console.WriteLine(new string(' ', Console.WindowWidth));
                     
+
 
                     if (totalEntries[i].FileType != "CONTENT")
                     {
@@ -184,36 +196,22 @@ namespace CriPakTools
                             {
                                 doTheThing = true;
                             }
-                            else if (((string)totalEntries[i].FileName).ToUpper() == replaceNames[j].Item1.ToUpper())
+                            else if (((string)totalEntries[i].FileName).ToUpper() == replaceNames[j].Item1)
                             {
                                 doTheThing = true;
                             }
 
                             if (doTheThing)
                             {
-                                byte[] newbie = File.ReadAllBytes(replaceNames[j].Item2);
+                                byte[] newbie = File.ReadAllBytes(replaceNames[i].Item2);
                                 totalEntries[i].FileSize = Convert.ChangeType(newbie.Length, totalEntries[i].FileSizeType);
                                 totalEntries[i].ExtractSize = Convert.ChangeType(newbie.Length, totalEntries[i].FileSizeType);
                                 cpk.UpdateFileEntry(totalEntries[i]);
                                 newCPK.Write(newbie);
 
-                                replacedFile = true;                                
+                                replacedFile = true;
 
-                                ConsoleColor prevColor = Console.ForegroundColor;
-                                Console.ForegroundColor = ConsoleColor.Cyan;
-
-                                int numLnClr = Console.CursorTop - consoleLine;
-                                Console.SetCursorPosition(0, consoleLine);
-
-                                Console.WriteLine(new string(' ', Console.WindowWidth * numLnClr));
-
-                                Console.SetCursorPosition(0, consoleLine);
-
-                                Console.WriteLine("Replaced file " + (++numReplaced) + " of " + replaceNames.Count + ": \"" + ((totalEntries[i].DirName != null) ? (string)totalEntries[i].DirName + "/" : "") + totalEntries[i].FileName.ToString().Trim() + "\".");
-
-                                consoleLine = Console.CursorTop;
-
-                                Console.ForegroundColor = prevColor;
+                                completedEntryReplacements.Add(totalEntries[i]);
 
                                 break;
                             }
@@ -253,8 +251,11 @@ namespace CriPakTools
                 newCPK.Close();
                 oldFile.Close();
 
-                File.Delete(cpk_name);
-                File.Move(cpk_name + ".tmp", cpk_name);
+                if (args.Length < 4)
+                {
+                    File.Delete(cpk_name);
+                    File.Move(cpk_name + ".tmp", cpk_name);
+                }
             }
             else if (selectedFunction == "INDEX" && args.Length == 3)
             {
